requirements.txt
Flask==3.0.0
pdfplumber==0.10.3
pandas==2.1.4
scikit-learn==1.3.2
python-dotenv==1.0.0
Werkzeug==3.0.1

# LangChain + Open Source LLM support
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
huggingface-hub>=0.20.0

# Optional: OpenAI (only if user wants to use OpenAI)
# openai>=1.6.1
# langchain-openai>=0.0.2